fix: remove scan parameter for nx cloud as deprecated
