+++ src/index.ts
@@ -1,38 +1,11 @@
 import * as core from '@actions/core';
 
 import { CommandBuilder } from './command-builder';
-import type { Inputs } from './inputs';
+import { parseInputs } from './inputs';
 import { runNx } from './run-nx';
 
-const ARGS_REGEX = /\w+|"(?:\\"|[^"])+"/g;
-
-function parseArgs(raw: string): string[] {
-  const parts = ARGS_REGEX.exec(raw);
-
-  if (!parts) return [];
-  return parts.filter((part) => part.length > 0);
-}
-
 async function main(): Promise<void> {
-  const inputs: Inputs = {
-    targets: core
-      .getInput('targets', { required: true })
-      .split(',')
-      .filter((target) => target.length > 0),
-    projects: core
-      .getInput('projects', { required: false })
-      .split(',')
-      .filter((project) => project.length > 0),
-    all: core.getInput('all') === 'true',
-    affected: core.getInput('affected') === 'true',
-    parallel:
-      parseInt(core.getInput('parallel')) === NaN
-        ? 3
-        : parseInt(core.getInput('parallel')),
-    args: parseArgs(core.getInput('args')),
-    nxCloud: core.getInput('nxCloud') === 'true',
-    workingDirectory: core.getInput('workingDirectory'),
-  };
+  const inputs = parseInputs();
 
   if (inputs.workingDirectory && inputs.workingDirectory.length > 0) {
     core.info(`🏃 Working in custom directory: ${inputs.workingDirectory}`);
